tests/wpt/web-platform-tests/referrer-policy/css-integration/child-css/processing-instruction.html

<!DOCTYPE html>
<html>
  <head>
    <title>CSS integration - child css via a ProcessingInstruction</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/common/utils.js"></script>
    <!-- Common global functions for referrer-policy tests. -->
    <script src="/referrer-policy/generic/common.js"></script>
    <meta name="referrer" content="origin">
  </head>
  <body>
    <p>Check that child css are loaded with the referrer and referrer policy the
    external stylesheet(referenced from a ProcessingInstruction).</p>
    <div class="styled"></div>

    <script>
      let css_test = async_test("Child css via a ProcessingInstruction.");
      let id = token();
      let url_prefix = location.protocol + "//www1." + location.hostname + ":" +
                       location.port +
                       "/referrer-policy/generic/subresource/stylesheet.py?id=" +
                       id;
      let css_url = url_prefix + "&amp;import-rule";
      let expected = url_prefix + "&import-rule";
      let check_url = url_prefix + "&report-headers";

      let processingInstruction =
        document.createProcessingInstruction(
          "xml-stylesheet", "href=\"" +css_url + "\" type=\"text/css\"");
      css_test.step_timeout(
          queryXhr.bind(this, check_url,
                        function(message) {
                          css_test.step(function() {
                            assert_own_property(message, "headers");
                            assert_own_property(message, "referrer");
                            assert_equals(message.referrer, expected);
                          });
                          css_test.done();
                        }),
          1000);
      document.insertBefore(processingInstruction, document.firstChild);
    </script>

    <div id="log"></div>
  </body>
</html>
